bonded into label with tooltip
